+++ test/eca/config_test.clj
@@ -1,0 +1,62 @@
+(ns eca.config-test
+  (:require
+   [clojure.test :refer [deftest is testing]]
+   [eca.config :as config]
+   [eca.test-helper :as h]
+   [matcher-combinators.matchers :as m]
+   [matcher-combinators.test :refer [match?]]))
+
+(h/reset-components-before-test)
+
+(deftest all-test
+  (testing "Default config"
+    (reset! config/initialization-config* {})
+    (is (match?
+         {:providers {"github-copilot" {:key m/absent
+                                        :models {"gpt-5" {}}}}}
+         (config/all {}))))
+  (testing "deep merging initializationOptions with initial config"
+    (reset! config/initialization-config* {:pureConfig true
+                                           :providers {"githubCopilot" {:key "123"}}})
+    (is (match?
+         {:pureConfig true
+          :providers {"github-copilot" {:key "123"
+                                        :models {"gpt-5" {}}}}}
+         (config/all {})))))
+
+(deftest deep-merge-test
+  (testing "basic merge"
+    (is (match?
+         {:a 1
+          :b 4
+          :c 3
+          :d 1}
+         (#'config/deep-merge {:a 1}
+                              {:b 2}
+                              {:c 3}
+                              {:b 4 :d 1}))))
+  (testing "deep merging"
+    (is (match?
+         {:a 1
+          :b {:c {:d 3
+                  :e 4}}}
+         (#'config/deep-merge {:a 1
+                               :b {:c {:d 3}}}
+                              {:b {:c {:e 4}}}))))
+  (testing "deep merging maps with other keys"
+    (is (match?
+         {:a 1
+          :b {:c {:e 3
+                  :f 4}
+              :d 2}}
+         (#'config/deep-merge {:a 1
+                               :b {:c {:e 3}
+                                   :d 2}}
+                              {:b {:c {:f 4}}})))
+    (is (match?
+         {:pureConfig true
+          :providers {"github-copilot" {:models {"gpt-5" {}}
+                                        :key "123"}}}
+         (#'config/deep-merge {:providers {"github-copilot" {:models {"gpt-5" {}}}}}
+                              {:pureConfig true
+                               :providers {"github-copilot" {:key "123"}}})))))